+++ CHANGELOG.md
@@ -1,10 +1,17 @@
 # Changelog
 
 - [Changelog](#changelog)
+  - [0.1.2](#012)
   - [0.1.1](#011)
   - [0.1.0](#010)
 
 ---
+
+## 0.1.2
+
+Released on 22/08/2022
+
+- New aphorisms
 
 ## 0.1.1